Redefine type sa_len_type as an alias of unsigned int.

+++ include/network/sa-len-type.h
@@ -18,7 +18,7 @@
 
 namespace Network
 {
-    using sa_len_type = unsigned short;
+    using sa_len_type = unsigned int;
 }
 
 #endif
